Handle mixed string and symbol results
Durning typical operation it is common to have a combination of cached
and rendered data. Cached data will usually have string keys, but
rendered data will usually have symbol keys. Combining keys now operates
on strings and uniqueness checks operate on the id key, whether that is
a string or a symbol.